+++ WhatToEat/Models/Data/ProductDTO.cs
@@ -15,7 +15,7 @@ namespace WhatToEat.Models.Data
        public string Name { get; set; }
        public string Slug { get; set; }
        public string Description { get; set; }
-       public decimal Calorie { get; set; }
+       public int Calorie { get; set; }
        public string CategoryName { get; set; }
        public int CategoryId { get; set; }
        public string ImageName { get; set; }
