+++ WhatToEat/Models/ViewModels/Diary/DiaryVM.cs
@@ -11,7 +11,7 @@ namespace WhatToEat.Models.ViewModels.Diary
         public int ProductId { get; set; }
         public string ProductName { get; set; }
         public int Quantity { get; set; }
-        public decimal Calorie { get; set; }
+        public int Calorie { get; set; }
         public decimal Total { get { return Quantity * Calorie; } }
         public string Image { get; set; }
     }
